requirements.txt
flask
flask-cors
openai
requests
python-dotenv
langchain
langchain-openai
